initial refactoring of gather operations
