+++ twister2/comms/src/java/edu/iu/dsc/tws/comms/dfw/io/gather/GatherBatchPartialReceiver.java
@@ -16,29 +16,16 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Queue;
-import java.util.concurrent.ArrayBlockingQueue;
 import java.util.logging.Logger;
 
-import edu.iu.dsc.tws.common.config.Config;
-import edu.iu.dsc.tws.comms.api.DataFlowOperation;
 import edu.iu.dsc.tws.comms.api.MessageFlags;
-import edu.iu.dsc.tws.comms.api.MessageReceiver;
-import edu.iu.dsc.tws.comms.dfw.ChannelMessage;
-import edu.iu.dsc.tws.comms.dfw.DataFlowContext;
 
-
-public class GatherBatchPartialReceiver implements MessageReceiver {
+public class GatherBatchPartialReceiver extends BaseGatherBatchReceiver {
   private static final Logger LOG = Logger.getLogger(GatherBatchPartialReceiver.class.getName());
 
   // lets keep track of the messages
   // for each task we need to keep track of incoming messages
-  private Map<Integer, Map<Integer, Queue<Object>>> messages = new HashMap<>();
-  private Map<Integer, Map<Integer, Integer>> counts = new HashMap<>();
-  private Map<Integer, Map<Integer, Boolean>> finished = new HashMap<>();
   protected Map<Integer, Boolean> isEmptySent = new HashMap<>();
-  private DataFlowOperation dataFlowOperation;
-  private int executor;
-  private int sendPendingMax = 128;
   private int destination;
   private Map<Integer, Boolean> batchDone = new HashMap<>();
 
@@ -46,65 +33,12 @@ public class GatherBatchPartialReceiver implements MessageReceiver {
     this.destination = dst;
   }
 
-
   @Override
-  public void init(Config cfg, DataFlowOperation op, Map<Integer, List<Integer>> expectedIds) {
-    executor = op.getTaskPlan().getThisExecutor();
-    sendPendingMax = DataFlowContext.sendPendingMax(cfg);
-
-    LOG.fine(String.format("%d gather partial expected ids %s", executor, expectedIds));
-    for (Map.Entry<Integer, List<Integer>> e : expectedIds.entrySet()) {
-      Map<Integer, Queue<Object>> messagesPerTask = new HashMap<>();
-      Map<Integer, Boolean> finishedPerTask = new HashMap<>();
-      Map<Integer, Integer> countsPerTask = new HashMap<>();
-
-      for (int i : e.getValue()) {
-        messagesPerTask.put(i, new ArrayBlockingQueue<>(sendPendingMax));
-        finishedPerTask.put(i, false);
-        countsPerTask.put(i, 0);
-      }
-      messages.put(e.getKey(), messagesPerTask);
-      finished.put(e.getKey(), finishedPerTask);
+  protected void init() {
+    for (Map.Entry<Integer, List<Integer>> e : expIds.entrySet()) {
       isEmptySent.put(e.getKey(), false);
-      counts.put(e.getKey(), countsPerTask);
       batchDone.put(e.getKey(), false);
     }
-    this.dataFlowOperation = op;
-    this.executor = dataFlowOperation.getTaskPlan().getThisExecutor();
-  }
-
-  @Override
-  public boolean onMessage(int source, int path, int target, int flags, Object object) {
-    // add the object to the map
-    boolean canAdd = true;
-
-    if (messages.get(target) == null) {
-      throw new RuntimeException(String.format("%d Partial receive error %d", executor, target));
-    }
-    Map<Integer, Boolean> finishedMessages = finished.get(target);
-    Queue<Object> m = messages.get(target).get(source);
-
-    if ((flags & MessageFlags.END) == MessageFlags.END) {
-      finishedMessages.put(source, true);
-      return true;
-    }
-
-
-    if (m.size() >= sendPendingMax) {
-      canAdd = false;
-    } else {
-      if (object instanceof ChannelMessage) {
-        ((ChannelMessage) object).incrementRefCount();
-      }
-      Integer c = counts.get(target).get(source);
-      counts.get(target).put(source, c + 1);
-
-      m.add(object);
-      if ((flags & MessageFlags.LAST) == MessageFlags.LAST) {
-        finishedMessages.put(source, true);
-      }
-    }
-    return canAdd;
   }
 
   @SuppressWarnings("unchecked")
